dynamic-optimize: align dynamic task layers and config DAGs
